pyproject.toml
[tool.poetry]
name = "lxc-manager"
version = "0.1.0"
description = ""
authors = ["Mathieu <mathieu.collet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
proxmoxer = "^2.0.1"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[tool.black]
line-length = 88
target-version = ['py38']
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | \.pytest_cache
  | \.mypy_cache
  | \.tox
  | \.nox
  | .*\.egg-info
  | \.dist-info
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503']
max-line-length = 88
max-complexity = 18
select = ['B','C','E','F','W','T4','B9']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
